enter tilemap manager
